requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
playwright==1.40.0
httpx==0.25.2
anthropic==0.34.0
pandas==2.1.3
PyPDF2==3.0.1
openpyxl==3.1.2
matplotlib==3.8.2
numpy==1.26.2
python-dotenv==1.0.0